+++ lib/app/presentation/app_root.dart
@@ -24,6 +24,7 @@ import 'package:zachranobed/common/presentation/notifiers/user_notifier.dart';
 import 'package:zachranobed/common/presentation/router/app_router.dart';
 import 'package:zachranobed/common/presentation/router/app_router.gr.dart';
 import 'package:zachranobed/common/presentation/utils/lifecycle_watcher.dart';
+import 'package:zachranobed/common/presentation/utils/orientation_lock.dart';
 import 'package:zachranobed/common/presentation/utils/ui_colors.dart';
 import 'package:zachranobed/common/presentation/utils/ui_text_styles.dart';
 import 'package:zachranobed/features/forceupdate/domain/usecase/check_if_upgrade_app_should_be_shown_usecase.dart';
@@ -253,9 +254,11 @@ class _AppRootState extends State<AppRoot> with LifecycleWatcher {
               return const SizedBox();
             }
 
-            return SoftUpdateWebBanner(
-              child: ConnectivityWrapper(
-                child: child,
+            return OrientationLock(
+              child: SoftUpdateWebBanner(
+                child: ConnectivityWrapper(
+                  child: child,
+                ),
               ),
             );
           },